requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
PyJWT==2.8.0
Werkzeug==3.0.1
pytest==7.4.3